Move music handling into its own module
